requirements.txt
streamlit
requests
python-dotenv
folium
streamlit-folium
pandas
openmeteo-requests
huggingface_hub
wikipedia
